优化 handlers 测试
