+++ crates/emmylua_ls/src/handlers/hover/test/hover_test.rs
@@ -1,10 +1,10 @@
 #[cfg(test)]
 mod tests {
 
-    use crate::handlers::hover::test::{HoverVirtualWorkspace, VirtualHoverResult};
+    use crate::handlers::test_lib::{ProviderVirtualWorkspace, VirtualHoverResult};
     #[test]
     fn test_1() {
-        let mut ws = HoverVirtualWorkspace::new();
+        let mut ws = ProviderVirtualWorkspace::new();
         assert!(ws.check_hover(
             r#"
                 ---@class <??>A
@@ -20,7 +20,7 @@ mod tests {
 
     #[test]
     fn test_right_to_left() {
-        let mut ws = HoverVirtualWorkspace::new();
+        let mut ws = ProviderVirtualWorkspace::new();
         assert!(ws.check_hover(
             r#"
                 ---@class H4
@@ -87,7 +87,7 @@ mod tests {
 
     #[test]
     fn test_hover_nil() {
-        let mut ws = HoverVirtualWorkspace::new();
+        let mut ws = ProviderVirtualWorkspace::new();
         assert!(ws.check_hover(
             r#"
                 ---@class A
@@ -106,7 +106,7 @@ mod tests {
 
     #[test]
     fn test_function_infer_return_val() {
-        let mut ws = HoverVirtualWorkspace::new();
+        let mut ws = ProviderVirtualWorkspace::new();
         assert!(ws.check_hover(
             r#"
                 local function <??>f(a, b)
